+++ CHANGELOG.md
@@ -1,8 +1,18 @@
 # v2.1.3
 
+Breaking changes:
+
+- Drops Rails 4 support
+- Drops Rails 5.0 support
+- Drops Rails 5.1 support
+- Drops Ruby 2.4 support
+
 Features:
 
-- Rails 6.1 compatibility
+- Adds Rails 6.1 support
+- Adds Ruby 2.7 support
+- Adds Ruby 3.0 support
+- Adds Ruby 3.1 support
 
 # v2.1.2